Change page for no waiting tutee infront of me
The no_waiting = True login in the views.py does not ever reach
since there will always be a waiting tutee log.
Therefore set the no_waiting to true not when there is no
waitingTutee(which is not possible) but when the length
of the waiting tutee list is equal to 1.
This means that the tutee that entered the room is the only one waiting
but the session has already started, so he still needs to wait
until the tutee before him finishes.